server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
spacy>=3.8.0
nltk>=3.9.0
scikit-learn>=1.6.0
pandas>=2.2.0
python-multipart>=0.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
PyPDF2>=3.0.0
boto3>=1.36.0
botocore>=1.36.0
